src/SeeJit.Core/Compilation/CSharpCompiler.cs
namespace SeeJit.Compilation
{
    using System.IO;
    using System.Linq;
    using System.Reflection;
    using Microsoft.CodeAnalysis;
    using Microsoft.CodeAnalysis.CSharp;

    public class CSharpCompiler
    {
        public static SemanticModel Compile(string assemblyName, string code, TextWriter errorWriter)
        {
            var syntaxTree = CSharpSyntaxTree.ParseText(code);

            var references = new MetadataReference[]
            {
                MetadataReference.CreateFromFile(typeof(object).Assembly.Location),
                MetadataReference.CreateFromFile(typeof(Enumerable).Assembly.Location)
            };

            var compilation = CSharpCompilation.Create(
                assemblyName,
                new[] { syntaxTree },
                references,
                new CSharpCompilationOptions(OutputKind.DynamicallyLinkedLibrary));

            using (var ms = new MemoryStream())
            {
                var result = compilation.Emit(ms);

                if (result.Success)
                {
                    Assembly.Load(ms.ToArray());

                    return compilation.GetSemanticModel(syntaxTree);
                }

                var failures = result.Diagnostics.Where(diagnostic =>
                    diagnostic.IsWarningAsError ||
                    diagnostic.Severity == DiagnosticSeverity.Error);

                foreach (var diagnostic in failures)
                {
                    errorWriter.WriteLine($"{diagnostic.Severity} {diagnostic.Id}: {diagnostic.GetMessage()}");
                }

                return null;
            }
        }
    }
}